// src/app/modules/user-management/model/employee.ts
export class Employee {
  id: number;
  employeeId: string;
  name: string;
  userName: string;
  passCode: string;
  role: string;
  status: string;
  accessStartDate: number;
  accessEndDate: number;
  contactNumber: number;
  emailId: string;
  address: string;
  description: string;
  lastLoggedIn: number;
  createdAt: number;
  createdBy: number;
  updatedAt: number;
  updatedBy: number;

  constructor(data?) {}
}